+++ 490-ce-commons/src/main/java/io/harness/ccm/commons/beans/Pricing.java
@@ -11,12 +11,10 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Value;
 
-import java.math.BigDecimal;
-
 @Value
 @Builder
 @AllArgsConstructor
 public class Pricing {
-  private BigDecimal pricePerHour;
+  private Double pricePerHour;
   private PricingSource source;
 }
